refactor(web): move media components into folders
