+++ web/src/components/ui/media/MediaReferenceTag.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { MediaTag } from "./MediaTag";
+import { MediaTag } from "../../MediaTag/MediaTag";
 import { useResolvedMedia } from "./useResolvedMedia";
 import { type MediaDescriptor } from "./mediaUtils";
 import { OBSERVATION_FIELD_SIZE_LIMIT_MEDIA_SOURCE } from "@langfuse/shared";
